code generation updates
Julia protoc codegen looks for option `generic_services` to be turned on for at least one of `c++`, python or Java. I think that is unnecessary, but until it is updated, added a check here to check and warn. Also added instructions to readme.

Package names with `.` are not handled well, needs some updates at protobuf code generation end. Have made the necessary changes needed here for now. And added instructions in readme to change package name to one without `.` before code generation.