set the TE header by default in requests
Some gRPC servers and proxies may need to be presented with a [`te` header](https://datatracker.ietf.org/doc/html/rfc7230#section-4.3). In gRPC trailers are always allowed in the protocol, but we encountered a scenario (#15) where the Python gRPC server implementation insists on it being [specified explicitly](https://nghttp2.org/blog/2015/03/24/proxying-grpc-with-nghttpx/).

This PR sets the TE header to `trailers` by default in every request.