concurrent use is flaky, add to readme
Concurrent use of gRPCClient is flaky, particularly on Julia >= 1.7.
The cause is unclear at this point. It could be something to do with
libcurl, or the way we use libcurl, or something specific to the way
libnghttp2 (that libcurl uses for HTTP2) works.

For the time being, I am adding this as a TODO item, and moving the
async tests under a flag.